refactor: DTO 정리 완료
ALL 패키지 이동 및 사용하지 않는 DTO 제거
