+++ src/main/java/wooteco/subway/admin/controller/LineStationController.java
@@ -3,8 +3,8 @@ package wooteco.subway.admin.controller;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.LineStationCreateResponse;
 import wooteco.subway.admin.dto.request.LineStationCreateRequest;
+import wooteco.subway.admin.dto.response.LineStationCreateResponse;
 import wooteco.subway.admin.dto.response.StationResponse;
 import wooteco.subway.admin.service.LineService;
 import wooteco.subway.admin.service.StationService;
